Add files via upload
🎉Creates a machine learning model that will make an appropriate description for a given image.

The backend takes in a list of image urls and returns a list of image descriptions. The model uses tensorflow to create a Recurrent Neural Network and a Convolutional Neural Network to analyze the image. The folder contains the weights of an already trained model in the model_data directory. The test.py file includes the function runModel() which will evaluate the images given.  Finally, the train_val.py can retrain the model. To retrain the model, you would have to download training images and the captions from Flickr8k dataset and update the config paths, both of which can be found in the readme. Update the config.py file if any paths change as well.