Updates to Readme, Removed pycache.
I updated the Readme.md based on the comments and it is now more applicable to our program. I removed the pycache because it was unneeded as well as a large file in model_data: captions.txt. I also added extra files that I realized weren't included that could be helpful to speed up testing.